Add support for CONSOLE_LOGGING_LEVEL environment variable in Node.js services
