+++ services/api/src/loggers/logger.ts
@@ -1,4 +1,5 @@
 const { addColors, createLogger, format, transports } = require('winston');
+import { getConfigFromEnv } from "../util/config";
 
 const config = {
   levels: {
@@ -42,7 +43,7 @@ export const logger = createLogger({
   ),
   transports: [
     new transports.Console({
-      level: 'debug',
+      level: getConfigFromEnv('CONSOLE_LOGGING_LEVEL', 'debug'),
       handleExceptions: true,
       json: false,
     }),
