+++ services/api/src/resources/deployment/helpers.ts
@@ -1,5 +1,5 @@
 import * as R from 'ramda';
-import { asyncPipe } from '@lagoon/commons/dist/util';
+import { asyncPipe } from '@lagoon/commons/dist/util/func';
 import { Pool } from 'mariadb';
 import { query } from '../../util/db';
 import { Sql } from './sql';
