+++ services/api/src/util/func.ts
@@ -1,10 +1,1 @@
-import { unless, is, isNil, partialRight, complement } from 'ramda';
-
-export const isNumber = is(Number);
-export const isArray = is(Array);
-
-export const toNumber = (input: string | number): number =>
-  unless(isNumber, partialRight(parseInt, [10]), input) as number;
-
-export const notArray = complement(isArray);
-export const isNotNil = complement(isNil);
+export * from '@lagoon/commons/dist/util/func';
